+++ src/extract.py
@@ -1,12 +1,14 @@
-from extraction.extractor import BigramExtractor as fe
+from extraction.bigram_extractor import extract
+# from extraction.flex_extractor import extract
+from utils.wrappers import export
 import sys
 
 if len(sys.argv) < 2:
     print("Usage: python3 extract.py path1 path2 ...")
     exit(1)
 
-ex = fe.BigramExtractor(sys.argv[1:])
+name = input("Name: ")
 print("extracting...")
-ex.extract()
+res = extract(sys.argv[1:])
 print("exporting...")
-ex.export("out.csv")
+export(res, name)
