+++ src/misc/intersection.py
